backend/requirements.txt
Django==3.2.23
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
django-cors-headers==4.3.1
pymongo==3.12.3
sqlparse==0.2.4
git+https://github.com/nesdis/djongo.git
torch==2.5.1
torchvision==0.20.1
Pillow==10.1.0
numpy==1.26.2
requests==2.31.0
python-decouple==3.8
APScheduler==3.11.1
sib-api-v3-sdk==7.6.0
bing-image-downloader==1.1.2